Added sent and received queries to db
