--- app/src/main/java/com/example/cs5520_project/RecylerViewAdapter.java
@@ -1,75 +1,0 @@
-package com.example.cs5520_project;
-
-import android.content.Context;
-import android.content.Intent;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.ImageView;
-import android.widget.TextView;
-
-import androidx.annotation.NonNull;
-import androidx.recyclerview.widget.RecyclerView;
-
-import java.util.ArrayList;
-
-public class RecylerViewAdapter extends RecyclerView.Adapter<RecylerViewAdapter.MyViewHolder> {
-
-    int []arr;
-    Context context;
-    String userName;
-    String friendName;
-
-    public RecylerViewAdapter(int[] arr, Context context, String userName, String friendName) {
-        this.arr = arr;
-        this.context = context;
-        this.userName = userName;
-        this.friendName = friendName;
-    }
-
-    @NonNull
-    @Override
-    public MyViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-        View view = LayoutInflater.from(parent.getContext()).inflate(R.layout.single_view, parent, false);
-        MyViewHolder myViewHolder = new MyViewHolder(view);
-
-        return myViewHolder;
-    }
-
-    @Override
-    public void onBindViewHolder(@NonNull MyViewHolder holder, int position) {
-        holder.imageView.setImageResource(arr[position]);
-        holder.textView.setText("Emoji " + (position + 1));
-        holder.clickedView.setText("Sent " + 0 + " times");
-    }
-
-    @Override
-    public int getItemCount() {
-        return arr.length;
-    }
-
-    public class MyViewHolder extends RecyclerView.ViewHolder implements View.OnClickListener {
-
-        ImageView imageView;
-        TextView textView;
-        TextView clickedView;
-        public MyViewHolder(@NonNull View itemView) {
-            super(itemView);
-            imageView = itemView.findViewById(R.id.imageView);
-            textView = itemView.findViewById(R.id.textView);
-            clickedView = itemView.findViewById(R.id.clickedView);
-            itemView.setOnClickListener(this);
-        }
-
-        @Override
-        public void onClick(View view) {
-            Intent intent = new Intent(context,StickerInfoActivity.class);
-            intent.putExtra("image",arr[getAbsoluteAdapterPosition()]);
-            intent.putExtra("title","Emoji " + (getAbsoluteAdapterPosition() + 1));
-            intent.putExtra("userName", userName);
-            intent.putExtra("friendName", friendName);
-            context.startActivity(intent);
-        }
-    }
-
-}
