+++ app/src/main/java/com/example/cs5520_project/UserInfo.java
@@ -6,7 +6,6 @@ public class UserInfo {
 
     String name, username, userId;
     HashMap<String, Integer> sentStickers, receivedStickers;
-    int numOfStickers;
 
     public UserInfo() {
     }
@@ -14,18 +13,12 @@ public class UserInfo {
     public UserInfo(String userId, String name, String username) {
         this.name = name;
         this.username = username;
-        this.numOfStickers = 0;
-       this.sentStickers =  new HashMap<String, Integer>();
-        this.receivedStickers =  new HashMap<String, Integer>();
-        sentStickers.put("test", 5);
-    }
-
-    public UserInfo(String userId, String name, String username, int numOfStickers) {
-        this.name = name;
-        this.username = username;
-        this.numOfStickers = numOfStickers;
         this.sentStickers =  new HashMap<String, Integer>();
         this.receivedStickers =  new HashMap<String, Integer>();
+        for(int i=1; i<11; i++){
+            sentStickers.put("Emoji " + i, 0);
+            receivedStickers.put("Emoji " + i, 0);
+        }
     }
 
     public String getName() {
@@ -42,14 +35,6 @@ public class UserInfo {
 
     public void setUsername(String username) {
         this.username = username;
-    }
-
-    public int getNumOfStickers() {
-        return 0;
-    }
-
-    public void setNumOfStickers(int numOfStickers) {
-        this.numOfStickers = numOfStickers;
     }
 
     public HashMap<String, Integer> getSentStickers() {
